snice:fix snice setting has no effect
Follow those steps
1、snice +4 -p ${pid}
2、ps -eo pid,args,nice | grep -w ${pid}
and snice did not work,the nice value still the same

This is because the read len is always equal to 128 bytes in
check_proc,so setpriority will never be called.

Here I delete equal branch,if the read len is greater than
sizeof(buf),then goto closure.