sysctl: set max or min values.
If the value to be set by sysctl is prefixed with < or >
it means only change the value if it needs to be hinger (or lower).

Like: sysctl -w 'fs.inotify:max_user_watches=>1024'

When packages installs their own settings in /etc/sysctl.d it may
cause conflicts between packages. This way this is resolved to the
minimal (or max) requirements for multiple packages.

Change-Id: If7d85603b2e7d8b84314c43991bf0540f38f543f