pkill: Add process_mrelease() support for immediate memory cleanup
process_mrelease() permits freeing a dying process' memory immediately
as part of the kill lifecycle, even if it has not yet fully exited. By
offloading the work of doing memory reclamation from the target process'
lifecycle and tying it to the the current pkill invocation, users can
free up resources more quickly than with a plain kill().

This is particularly useful when terminating processes that linger in
uninterruptible sleep or otherwise delay when presented with terminal
signals. With a plain kill(), we can contend with the inherent delay of
the target process cleaning up its own memory (even on SIGKILL, as we
may need to wait for the return to usermode in cases like DMA or some
I/O). Invoking process_mrelease allows the caller to immediately reclaim
the process' memory and make its memory available to the system more
quickly.

Signed-off-by: Chris Down <chris@chrisdown.name>